Moved caching functionality from ONEDrive to Workspace and reimplemented it so that caching of all information that is available in the Solr search index is downloaded and cached when the workspace is initially opened.
Implemented a system in the Workspace that allows its caching system to also
be used for objects that are not in any of the Workspace folders. This is
used by the ONEDrive FlatSpace folder.

The Workspace also handles downloading and caching of sciobj and sysmeta.

All configuration settings have defaults in settings.py. These can be overriden
when creating the Workspace object by passing in a dictionary of key/values
where the keys match the ones in settings.py.

Updated Workspace types to PyXB v1.2.3 bindings.