Removed backwards compatibility in RESTClient
RESTClient was using conditional logic that checked the types of
parameters in order to support older client code. This removes that and
works with current client code that has been updated correspondingly. In
the unlikely case that any 3rd party code uses RESTClient directly, it
may have to be updated as well.