Add truncating of long DataONEExceptions strings
DataONEExceptions can sometimes contain very long documents, such as
when  HTML documents returned by servers are added as trace information
for diagnostics. To avoid dumping large documents to logs when
serializing to strings for logging, now truncates individual fields that
are longer than 1K when serializing the full exception to string. The
full, untruncated, values are still returned when accessing the fields
in exception objects.