Add v2 APIs with SID support
Also:
- Refactor to create separate layers of responsibility between the views
and GMN internal functionality.
- Refactor to match model changes required by newer versions of Django
and changes to support SID/PID uniqueness.
- Improved support for obsolescence chains both in v1 and v2.
- The old version of this file was called v1.py because it was expected
that there would be a v2.py for the v2 API. However, there was so much
duplication of code between the APIs that it was better to support both
APIs with the same views and using a decorator to resolve SIDs.