Add management command for migrating from v1 to v2 database
The main strategy is scanning through the SysMeta XML docs of all the
SciObjs stored in the v1 instance and creating representations for them
in the v2 database. Information not stored in the XML docs is pulled
directly from the v1 database.

Includes a safety function to prevent overwriting any existing
information in the v2 database.