Add decorator that decodes the SID or PID extracted from URL path segment
Add decorator that decodes the SID or PID extracted from URL path
segment  by Django.

Currently, Django passes percent-encoded params to views when they
were extracted from URL path segments by the Django URL regex parser and
dispatcher. IMO, that's a bug and I'm working with Django devs to see if
this can be fixed. For now, we need this.