Improved handling of replicas that are in obsolescence chains
- This fixes an issue where GMN could not hold replicas that were in
obsolescence chains due to the way the database was layed out and the
constraints that had been put in place. GMN now keeps track of PIDs
referenced by replias and reserves them from local use, preventing a
situation where someone could cause unrelated objects to appear to be in
a chain.
- A secondary issue has been fixed, where accepted but not yet processed
replicas were not counting towards the configured space allocation for
replicas. This could cause GMN v1 to accept replicas for which there
would be no space after processing.