Remove unhelpful unit tests
Sadly, had to remove 6000 lines of unit tests that tested irrelevant
internal implementation details and did not indicate if the methods
being tested worked as expected.

Almost all of the tests were on the form, "call external method A and
assert that it in turn calls internal method B with parameter C". So,
the success or failure of the test did not indicate if the tested method
worked.

These unit tests also locked down our internal implementation details
and caused unit testing failures whenever something was refactored, even
though the methods being tested actually still worked.

In particular, the refactoring associated with the move to Requests and
the new Session class broke many of these tests. We're now testing with
Responses instead.