Add tool to verify that Wheel package includes tracked files
Verify that Wheel package includes the exact set of files tracked by Git

This is a workaround for the way setuptools works, which is basically
that it vacuums up everything that looks like a Python script in
anything that looks like a package, which makes it easy to publish local
"junk" files by accident.

More notes in the source.