Modify instance generator to enable testing of shared resources
The original instance generator generated subjects, identifiers, node
references, etc, of random strings that were so long as to be
practically always unique. This caused e.g., the ability of the MN to
hold two objects that had the same submitter to be untested. The new
instance generator by default provides the ability to simulate a D1 env
where there will be both shared and unshared resources.